* src/update-notifier.c (check_for_updates): Cleaned up a warning 	on compilation. 	(xexp_list_sort): Likewise.
	* src/mime-open.c: Added two required includes, to avoid
	compilation warnings.

	* src/mime-server.c: Added two required includes, to avoid
	compilation warnings.

	* src/apt-worker-bin.cc (substreq): Cleaned up a warning on
	compilation.

	* utils/maemo-select-menu-location-main.c
	(selection_location_response): Cleaned up a warning on
	compilation.



git-svn-id: https://garage.maemo.org/svn/hildon-app-mgr/trunk@892 e55152ac-3a23-0410-980c-f88739b87167

+++ src/apt-worker-bin.cc
@@ -4547,7 +4547,7 @@ encode_localized_field (pkgTagSection *section, const char *field,
 static bool
 substreq (const char *start, const char *end, const char *str)
 {
-  return end-start == strlen (str) && !strncmp (start, str, end-start);
+  return ((size_t)(end-start) == strlen (str)) && !strncmp (start, str, end-start);
 }
 
 static int
